refactor(insights): adopt newspack Wizard as page shell
Replaces the custom Insights header, tab navigation, and tab routing
with the design-system Wizard + TabbedNavigation so the page chrome
matches the rest of the Newspack admin. Tab URLs switch from ?tab=X
to #/X with a one-shot redirect on mount for legacy links. The
DateRangePicker and ComparisonToggle render above tab content via
the Wizard's renderAboveSections slot.

This is a reversible structural change - see git revert <this-sha>
if visual or behavioral feedback warrants undoing it.
