+++ plugins/newspack-plugin/src/wizards/insights/style.scss
@@ -1,10 +1,11 @@
 /**
  * Newspack Insights wizard chrome (NPPD-1602)
  *
- * Per spec at ~/Sites/insights-docs/component-design-spec.md. Owns
- * page-level layout (header, tab nav, tab content area) plus styles for
- * the chrome-only components (DateRangePicker, ComparisonToggle,
- * LastUpdated, tab stub).
+ * Page shell (admin header, tab navigation, footer, content max-width)
+ * is owned by the design-system Wizard component. This file scopes
+ * Insights-only chrome: the renderAboveSections control row
+ * (DateRangePicker + ComparisonToggle), the empty state, LastUpdated,
+ * tab fallback, tab stub, and section header containers.
  *
  * Also forwards the shared cross-tab chrome — sections, metric cards,
  * tables, tab loading/error — from `tabs/components/sections.scss` so
@@ -18,38 +19,22 @@
 @use "tabs/components/sections";
 
 .newspack-insights {
-	max-width: 1200px;
-	margin: 0 auto;
-	padding: 24px 24px 64px;
 	color: wp-colors.$gray-900;
 
-	&__header {
-		display: flex;
-		flex-wrap: wrap;
-		justify-content: space-between;
-		align-items: flex-start;
-		gap: 16px;
-		margin-bottom: 24px;
-	}
-
-	&__header-left {
-		display: flex;
-		flex-direction: column;
-		gap: 4px;
-	}
-
-	&__header-right {
+	// Date range picker + comparison toggle row rendered via the
+	// Wizard's renderAboveSections slot. Sits above the section
+	// content, right-aligned to mirror the old header-right placement.
+	&__header-controls {
 		display: flex;
 		flex-wrap: wrap;
 		align-items: center;
+		justify-content: flex-end;
 		gap: 16px;
-	}
+		margin-bottom: 24px;
 
-	&__title {
-		font-size: 28px;
-		font-weight: 600;
-		line-height: 1.2;
-		margin: 0;
+		@media ( max-width: 782px ) {
+			justify-content: flex-start;
+		}
 	}
 
 	&__last-updated {
@@ -150,49 +135,7 @@
 		font-size: 13px;
 	}
 
-	// Tab navigation
-
-	&__tabs {
-		display: flex;
-		flex-wrap: wrap;
-		gap: 0;
-		margin-bottom: 24px;
-		border-bottom: 1px solid wp-colors.$gray-200;
-	}
-
-	&__tab {
-		appearance: none;
-		background: transparent;
-		border: 0;
-		padding: 12px 16px;
-		margin: 0 0 -1px 0;
-		font-size: 13px;
-		font-weight: 500;
-		color: wp-colors.$gray-700;
-		cursor: pointer;
-		border-bottom: 2px solid transparent;
-		transition: color 120ms ease, border-color 120ms ease;
-
-		&:hover {
-			color: wp-colors.$gray-900;
-		}
-
-		&:focus-visible {
-			outline: 2px solid var(--wp-admin-theme-color);
-			outline-offset: -2px;
-		}
-
-		&.is-active {
-			color: var(--wp-admin-theme-color);
-			border-bottom-color: var(--wp-admin-theme-color);
-		}
-	}
-
 	// Tab content area
-
-	&__tab-content {
-		min-height: 320px;
-	}
 
 	&__tab-fallback {
 		display: flex;
